# gen/openapi/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "REST API of the codebase-go service template. Every response uses the same JSON envelope: {data, meta} on success and {error, meta} on failure.",
        "title": "codebase-go API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/v1/users": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Registers a user and publishes the user.created event (when MQ is enabled).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Create a user",
                "operationId": "createUser",
                "parameters": [
                    {
                        "description": "User to create",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/rest.DataResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.UserResponse"
                                        }
                                    }
                                }
                            ]
                        },
                        "headers": {
                            "Location": {
                                "type": "string",
                                "description": "URL of the created user"
                            }
                        }
                    },
                    "400": {
                        "description": "Validation failed (see error.details)",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Email already registered",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/users/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Reads from Redis first (cache-aside), then PostgreSQL.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get a user by ID",
                "operationId": "getUser",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "User ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/rest.DataResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.UserResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/healthz": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Liveness probe",
                "operationId": "liveness",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.HealthResponse"
                        }
                    }
                }
            }
        },
        "/readyz": {
            "get": {
                "description": "Checks every enabled dependency (PostgreSQL, Redis, RabbitMQ).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Readiness probe",
                "operationId": "readiness",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.HealthResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/rest.HealthResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.CreateUserRequest": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "format": "email",
                    "maxLength": 255,
                    "example": "andi@example.com"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2,
                    "example": "Andi Pratama"
                },
                "password": {
                    "description": "bcrypt limit is 72 bytes",
                    "type": "string",
                    "maxLength": 72,
                    "minLength": 8,
                    "example": "Password123!"
                },
                "phone": {
                    "type": "string",
                    "example": "+6281234567890"
                }
            }
        },
        "dto.UserResponse": {
            "type": "object",
            "required": [
                "created_at",
                "email",
                "id",
                "name",
                "phone",
                "updated_at"
            ],
            "properties": {
                "created_at": {
                    "type": "string",
                    "format": "date-time"
                },
                "email": {
                    "type": "string",
                    "example": "andi@example.com"
                },
                "id": {
                    "type": "string",
                    "format": "uuid",
                    "example": "01928f6a-0000-7000-8000-000000000001"
                },
                "name": {
                    "type": "string",
                    "example": "Andi Pratama"
                },
                "phone": {
                    "type": "string",
                    "example": "+6281200000001"
                },
                "updated_at": {
                    "type": "string",
                    "format": "date-time"
                }
            }
        },
        "model.FieldError": {
            "type": "object",
            "required": [
                "field",
                "message"
            ],
            "properties": {
                "field": {
                    "type": "string",
                    "example": "email"
                },
                "message": {
                    "type": "string",
                    "example": "must be a valid email address"
                }
            }
        },
        "rest.DataResponse": {
            "type": "object",
            "required": [
                "data",
                "meta"
            ],
            "properties": {
                "data": {},
                "meta": {
                    "$ref": "#/definitions/rest.Meta"
                }
            }
        },
        "rest.ErrorBody": {
            "type": "object",
            "required": [
                "code",
                "message"
            ],
            "properties": {
                "code": {
                    "type": "string",
                    "example": "NOT_FOUND"
                },
                "details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.FieldError"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "not found: user 01928f6a-0000-7000-8000-000000000001"
                }
            }
        },
        "rest.ErrorResponse": {
            "type": "object",
            "required": [
                "error",
                "meta"
            ],
            "properties": {
                "error": {
                    "$ref": "#/definitions/rest.ErrorBody"
                },
                "meta": {
                    "$ref": "#/definitions/rest.Meta"
                }
            }
        },
        "rest.HealthResponse": {
            "type": "object",
            "properties": {
                "checks": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "status": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
        "rest.Meta": {
            "type": "object",
            "properties": {
                "request_id": {
                    "type": "string",
                    "example": "6f1c2b9e-1d2a-4a51-9f3e-2b7c1d0e8a11"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Access token. Format: \"Bearer \u003ctoken\u003e\".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}